handling exception where is no scheduled actions

+++ lib/ecs_autoscaling_scheduler/cli/destroy.rb
@@ -6,9 +6,16 @@ module EcsAutoscalingScheduler
   class Cli
     class Destroy
       def run
-        cluster_name          = ask_cluster_name
-        service_name          = ask_service_name(cluster_name)
-        scheduled_action_name = ask_scheduled_action_name(cluster_name, service_name)
+        cluster_name = ask_cluster_name
+        service_name = ask_service_name(cluster_name)
+
+        scheduled_action_names = application_auto_scaling_client.describe_scheduled_actions(cluster_name: cluster_name, service_name: service_name).map(&:scheduled_action_name)
+        if scheduled_action_names.length == 0
+          puts "There is no scheduled action."
+          return
+        end
+
+        scheduled_action_name = ask_scheduled_action_name(scheduled_action_names)
 
         if ask_ok
           application_auto_scaling_client.delete_scheduled_action(
@@ -43,8 +50,8 @@ module EcsAutoscalingScheduler
           prompt.select("Which service do you want to manage?", ecs_client.all_service_names(cluster: cluster), required: true)
         end
 
-        def ask_scheduled_action_name(cluster_name, service_name)
-          prompt.select("Which scheduled action do you want to destroy?", application_auto_scaling_client.describe_scheduled_actions(cluster_name: cluster_name, service_name: service_name).map(&:scheduled_action_name), required: true)
+        def ask_scheduled_action_name(scheduled_action_names)
+          prompt.select("Which scheduled action do you want to destroy?", scheduled_action_names, required: true)
         end
 
         def ask_ok
